wafflow/bookmark/tests.py
from django.contrib.auth.models import User
from django.test import TestCase, Client
from rest_framework import status
from rest_framework.authtoken.models import Token

from answer.models import Answer, UserAnswer
from comment.models import Comment
from question.models import Question
from user.models import UserProfile

import json
