requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
cfgv==3.2.0
click==7.1.2
distlib==0.3.1
Django==3.1.4
djangorestframework==3.12.2
filelock==3.0.12
identify==1.5.10
mypy-extensions==0.4.3
nodeenv==1.5.0
pathspec==0.8.1
pre-commit==2.9.3
psycopg2-binary==2.8.6
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
virtualenv==20.2.2
